Adjust signature of fillOptionalFootLocationCloud to float,float instead of double,double to prevent narrowing conversion

+++ ros1/src/data/march_realsense_reader/include/pointcloud_processor/parameter_determiner.h
@@ -136,7 +136,7 @@ protected:
 
     // Fill the foot locations to try cloud with a line of points from (start,
     // 0) to (end, 0)
-    bool fillOptionalFootLocationCloud(double start, double end);
+    bool fillOptionalFootLocationCloud(float start, float end);
 
     // Read all relevant parameters from the parameter yaml file
     void readYaml();
